requirements.txt
bcrypt==4.0.1
channels==3.0.4
django==4.0.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
djongo==1.3.6
drf-spectacular==0.27.1
drf-yasg==1.21.7
matplotlib==3.7.2
mysqlclient==2.2.1
numpy==1.26.0
opencv-contrib-python==4.8.1.78
pandas==2.1.0
pytz==2023.3
pymongo==3.12.1
pymysql==1.1.0
keras==2.14.0
tensorflow==2.14.0
spacy==3.6.1
pyvi==0.1.1
https://gitlab.com/trungtv/vi_spacy/-/raw/master/packages/vi_core_news_lg-3.6.0/dist/vi_core_news_lg-3.6.0.tar.gz